Rearrange source code Add login page
