--- src/HelpPage.css
@@ -1,135 +1,0 @@
-.help-page {
-    padding: 20px;
-    font-family: 'Roboto', sans-serif;
-    overflow-y: auto;
-}
-
-.help-page::-webkit-scrollbar {
-    width: 10px;
-}
-
-.help-page::-webkit-scrollbar-thumb {
-    background-color: #bbb;
-    border-radius: 8px;
-}
-
-.help-page::-webkit-scrollbar-thumb:hover {
-    background-color: #aaa;
-}
-
-.help-page::-webkit-scrollbar-track {
-    background-color: #fff;
-}
-
-h1 {
-    font-size: 28px;
-    color: #e48b3d;
-    margin-bottom: 20px;
-}
-
-.help-container {
-    display: flex;
-    flex-direction: column;  /* Ensures items are arranged vertically */
-    justify-content: flex-end;
-    gap: 40px;
-}
-
-.accordion {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    gap: 20px;
-}
-
-.accordion-item {
-    background-color: #fff;
-    padding: 20px;
-    border: 1px solid #ddd;
-    border-radius: 8px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-    cursor: pointer;
-}
-
-.accordion-item h3 {
-    font-size: 20px;
-    margin: 0;
-    color: #333;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding-right: 20px; /* Ensure space for the icon */
-}
-
-.accordion-content {
-    margin-top: 10px;
-    font-size: 16px;
-    color: #666;
-    display: none; /* Start hidden */
-    transition: all 0.3s ease-in-out;
-}
-
-.accordion-item.active .accordion-content {
-    display: block; /* Show content when active */
-}
-
-.accordion-item.active {
-    background-color: #fafafa;
-}
-
-.accordion-item h3:after {
-    content: '\02795'; /* Unicode character for "plus" sign (+) */
-    font-size: 20px; /* Increase font size for better visibility */
-    color: #777;
-}
-
-.accordion-item.active h3:after {
-    content: "\2796"; /* Unicode character for "minus" sign (-) */
-    color: #e48b3d; /* Change color when active */
-}
-
-.contact-info {
-    flex: 1;
-    background-color: #f9f9f9;
-    padding: 20px;
-    border-radius: 8px;
-    border: 1px solid #ddd;
-}
-
-.contact-info h2 {
-    font-size: 24px;
-    color: #e48b3d;
-}
-
-.contact-info p {
-    font-size: 16px;
-    color: #555;
-}
-
-.contact-info a {
-    color: #e48b3d;
-    text-decoration: none;
-}
-
-.contact-info a:hover {
-    text-decoration: underline;
-}
-
-.faq {
-    margin-top: 40px;
-}
-
-.faq h3 {
-    font-size: 20px;
-    color: #333;
-}
-
-.faq ul {
-    list-style-type: none;
-    padding: 0;
-}
-
-.faq ul li {
-    font-size: 16px;
-    color: #555;
-    margin-bottom: 10px;
-}
