+++ src/app/DetailsPage.css
