+++ src/app/DetailsPage.js
