+++ src/app/Header.css
@@ -1,0 +1,97 @@
+/* Header Container */
+.header {
+    display: flex; /* Horizontal layout */
+    justify-content: space-between; /* Space between nav and buttons */
+    align-items: center; /* Align items vertically */
+    padding: 10px 20px;
+    height: 70px; /* Set a fixed height for the header */
+    background-color: #fdf2e9;
+    border-bottom: 1px solid #e0e0e0;
+    box-shadow: 0px 2px 5px rgba(0, 0, 0, 0.1);
+    position: sticky;
+    top: 0;
+    z-index: 1000;
+}
+
+/* Logo and text*/
+.logo-container {
+    display: flex;
+    align-items: center;
+    gap: 1px;
+    cursor: pointer;
+    height: 100%;
+    overflow: hidden;
+}
+
+.header-logo {
+    max-height: 100%;
+    width: auto;
+    object-fit: contain;
+}
+
+.header-text {
+    max-height: 250%;
+    width: auto;
+    object-fit: contain;
+    margin-top: 40px;
+    margin-left: -50px;
+    padding: 0;
+}
+
+/* Navigation Links */
+.nav-links {
+    display: flex;
+    gap: 30px;
+    align-items: center;
+    margin-right: auto;
+}
+
+.nav-links a {
+    text-decoration: none;
+    font-size: 16px;
+    color: #555;
+    transition: color 0.3s ease;
+    cursor: pointer;
+}
+
+.nav-links a:hover {
+    color: #e48b3d;
+}
+
+/* Buttons on the Right */
+.header-buttons {
+    display: flex;
+    gap: 10px;
+}
+
+.search-button,
+.upload-button {
+    padding: 8px 16px;
+    border: none;
+    gap: 8px;
+    border-radius: 6px;
+    font-size: 14px;
+    font-weight: bold;
+    cursor: pointer;
+    transition: background-color 0.3s ease, transform 0.2s ease;
+}
+
+.search-button {
+    background-color: #e48b3d;
+    color: #fff;
+}
+
+.search-button:hover {
+    background-color: #d9792c;
+    color: #fff;
+}
+
+.upload-button {
+    background-color: #e48b3d;
+    color: #fff;
+}
+
+.upload-button:hover {
+    background-color: #d9792c;
+    color: #fff;
+}
