+++ src/app/UserPage.css
@@ -1,0 +1,180 @@
+.user-page {
+    padding: 20px;
+    font-family: 'Roboto', sans-serif;
+    overflow-y: auto;
+}
+
+.user-page::-webkit-scrollbar {
+    width: 10px;
+}
+
+.user-page::-webkit-scrollbar-thumb {
+    background-color: #bbb;
+    border-radius: 8px;
+}
+
+.user-page::-webkit-scrollbar-thumb:hover {
+    background-color: #aaa;
+}
+
+.user-page::-webkit-scrollbar-track {
+    background-color: #fff;
+}
+
+h1 {
+    font-size: 28px;
+    color: #e48b3d;
+    margin-bottom: 20px;
+    text-align: center; /* 居中显示用户名 */
+}
+
+.profile-container {
+    display: flex;
+    gap: 40px;
+}
+
+/* Profile Information Section */
+.profile-info {
+    flex: 1;
+    background-color: #fff;
+    padding: 20px;
+    border-radius: 8px;
+    border: 1px solid #ddd;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.profile-info h2 {
+    font-size: 24px;
+    color: #333;
+    margin-bottom: 20px;
+}
+
+.profile-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.edit-button {
+    background-color: #e48b3d;
+    color: white;
+    border: none;
+    padding: 8px 12px;
+    border-radius: 6px;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+}
+
+.edit-button:hover {
+    background-color: #d9792c;
+}
+
+.profile-fields {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+}
+
+.profile-fields div {
+    display: flex;
+    justify-content: space-between;
+    font-size: 16px;
+    color: #555;
+}
+
+.profile-fields label {
+    font-weight: bold;
+    color: #333;
+}
+
+.profile-fields input {
+    padding: 5px 10px;
+    font-size: 14px;
+    border: 1px solid #ddd;
+    border-radius: 4px;
+    width: 70%;
+}
+
+.profile-fields input[type="date"] {
+    width: 50%;
+}
+
+.profile-fields a {
+    color: #e48b3d;
+    text-decoration: none;
+}
+
+.profile-fields a:hover {
+    text-decoration: underline;
+}
+
+.action-buttons {
+    margin-top: 20px;
+    display: flex;
+    gap: 10px;
+}
+
+.save-button {
+    background-color: #4CAF50;
+    color: white;
+    border: none;
+    padding: 8px 12px;
+    border-radius: 6px;
+    cursor: pointer;
+}
+
+.cancel-button {
+    background-color: #f44336;
+    color: white;
+    border: none;
+    padding: 8px 12px;
+    border-radius: 6px;
+    cursor: pointer;
+}
+
+/* Favourites Section */
+.favourites {
+    flex: 2;
+    background-color: #fff;
+    padding: 20px;
+    border-radius: 8px;
+    border: 1px solid #ddd;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.favourites h2 {
+    font-size: 24px;
+    color: #333;
+    margin-bottom: 20px;
+}
+
+.favourites-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(150px, 1fr));
+    gap: 20px;
+}
+
+.favourite-card {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 10px;
+    border: 1px solid #ddd;
+    border-radius: 4px;
+    background-color: #fff;
+    transition: transform 0.3s ease;
+    cursor: pointer;
+}
+
+.favourite-card:hover {
+    transform: scale(1.05);
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+}
+
+.image-placeholder {
+    background-color: #e0e0e0;
+    width: 100px;
+    height: 100px;
+    margin-bottom: 10px;
+    border-radius: 4px;
+}
