+++ src/auth/ProtectedRoute.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { useLocation, Navigate } from 'react-router-dom';
+import { useAuth } from './AuthContext';
+
+const ProtectedRoute = ({ children }) => {
+    const { isAuthenticated } = useAuth();
+    const location = useLocation();
+
+    return isAuthenticated ? (
+        children
+    ) : (
+        <Navigate to="/login" state={{ parent: location }} />
+    );
+};
+
+export default ProtectedRoute;
